Moved RelatableView into its own folder
