+++ src/browser/modules/Stream/CypherFrame/RelatableView/relatable-view.test.tsx
